RFC-004 P3: skill dispatch extracted to native_tools/skills_tools.py
The skill CRUD/meta/invoke/dynamic dispatch — ~120 lines of inline
if/elif living inside NativeToolDispatcher.dispatch() — moves verbatim
to a SkillTools domain owner, completing the native side's own
decomposition (the 6th domain that RFC-002 left embedded in the table
file).

- SkillTools built by the dispatcher on the SAME objects (shared
  references), so monkeypatching dispatcher.skill_manager reaches both
- effects passed INTO skills.dispatch (constructed by the registry) so
  skills_tools never imports registry — no cycle
- dispatcher.handles() = table check + skills.handles() (same truth
  table, pinned by P0)
- SKILL_CRUD_TOOLS re-exported from registry for existing importers
- the two file-callback pins re-pointed at the new owner (d.skills),
  per the plan's per-wave re-pointing rule
- native registry.py 324 -> 193 lines; both files at zero ruff findings

Behavior pinned by test_native_skill_dispatch_pins.py — all green
unmodified except the two re-pointed callback pins. Full suite 6,123.
